- Added String concatenation case in invokedynamic instrumentation and HeapVM (In progress) - Added DSESystemTestBase for general testing and invokedynamic cases
